+++ src/main/java/com/chatbot/unla/util/ListarUsuariosPdf.java
@@ -119,7 +119,7 @@ public class ListarUsuariosPdf extends AbstractPdfView {
 			celda.setPadding(5);
 			tablaLista.addCell(celda);
 			
-			Integer documento = usuario.getDocumento();
+			Long documento = usuario.getDocumento();
 			celda = new PdfPCell(new Phrase(documento.toString(), fuenteDatos));
 			celda.setPadding(5);
 			tablaLista.addCell(celda);
